Fix modules splitting issues from rebase
